requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
cloudscraper
cachetools
httpx
sqlalchemy
aiofiles
